Uploaded server and some other data
I've added the server.jar file plus the server_data. This also includes the mail.properties and the server.properties files, which are both used to configure the server. I had some issues sending myself the Auth Token, but all of those issues are fixed in this final code and settings. Just change my email addresses to yours